chore: rename storybook mock file
So it doesn't get picked up by jest during tests
