+++ app/frontend/src/features/auth/password/CompleteResetPassword.stories.tsx
@@ -3,7 +3,7 @@ import { Empty } from "google-protobuf/google/protobuf/empty_pb";
 import React from "react";
 import { MemoryRouter, Route } from "react-router-dom";
 import { resetPasswordRoute } from "routes";
-import { mockedService } from "stories/__mocks__/service";
+import { mockedService } from "stories/serviceMocks";
 
 import CompleteResetPassword from "./CompleteResetPassword";
