+++ app/frontend/src/features/communities/discussion/DiscussionsListPage.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, Story } from "@storybook/react";
-import { mockedService } from "stories/__mocks__/service";
+import { mockedService } from "stories/serviceMocks";
 import community from "test/fixtures/community.json";
 import discussions from "test/fixtures/discussions.json";
